+++ RetroBat/RetroBat/IniReader.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
@@ -51,7 +51,7 @@ namespace RetroBat
 LanguageDetection=1
 
 ; At startup RetroBat will reset the default config files options of emulationstation and retrobat.ini.
-; Use at your own risk.	
+; Use at your own risk.
 ResetConfigMode=0
 
 ; Run automatically RetroBat at Windows startup (0=NO 1=STARTUP 2=REGISTRY).
@@ -111,7 +111,7 @@ FocusDelay=2000
 ; The frontend will parse only the gamelist.xml files in roms directories to display available games.
 ; If files are added when this option is enabled, they will not appear in the gamelists of the frontend. The option must be enabled again to display new entries properly.
 GameListOnly=0
- 
+
 ; 0 = run the frontend normally.
 ; 1 = run the frontend in kiosk mode.
 ; 2 = run the frontend in kid mode.
@@ -239,7 +239,8 @@ DrawFramerate=0";
             }
             catch (Exception ex)
             {
-                throw ex;
+                SimpleLogger.Instance.Error("[IniFile] Failed to parse ini file " + path, ex);
+                throw;
             }
         }
 
@@ -696,7 +697,7 @@ DrawFramerate=0";
     "--draw-framerate		display the framerate"
     "--no-exit			don't show the exit option in the menu"
     "--no-splash			don't show the splash screen"
-    "--debug				more logging, show console on Windows"				
+    "--debug				more logging, show console on Windows"
     "--windowed			not fullscreen, should be used with --resolution"
     "--vsync [1/on or 0/off]		turn vsync on or off (default is on)"
     "--max-vram [size]		Max VRAM to use in Mb before swapping. 0 for unlimited"
