+++ RetroBat/RetroBat/RetroBat.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -58,16 +58,19 @@
     <Reference Include="WindowsFormsIntegration" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AutostartManager.cs" />
+    <Compile Include="DpiAwarenessManager.cs" />
+    <Compile Include="EmulationStationLauncher.cs" />
+    <Compile Include="EmulationStationSettings.cs" />
     <Compile Include="FocusHelper.cs" />
     <Compile Include="IniReader.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RawInput.cs">
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Include="RegistryTools.cs" />
+    <Compile Include="RetroBatConfigLoader.cs" />
     <Compile Include="SimpleLogger.cs" />
     <Compile Include="SplashVideo.cs" />
+    <Compile Include="StartupFileChecker.cs" />
     <Compile Include="VideoPlayerForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -75,7 +78,6 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="app.manifest" />
-    <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="resources\launcher.ico" />
